release 0.0.86 source code for go

+++ services/functiongraph/v2/model/model_sla_reports_value.go
@@ -9,10 +9,10 @@ import (
 type SlaReportsValue struct {
 
 	// 时间戳
-	Timestamp *int32 `json:"timestamp,omitempty"`
+	Timestamp *int64 `json:"timestamp,omitempty"`
 
 	// 值
-	Value *int32 `json:"value,omitempty"`
+	Value *float64 `json:"value,omitempty"`
 }
 
 func (o SlaReportsValue) String() string {
